+++ Citadel/src/utopia/citadel/database/access/many/device/ManyClientDeviceUsersAccess.scala
@@ -25,6 +25,7 @@ object ManyClientDeviceUsersAccess
   * @author Mikko Hilpinen
   * @since 2021-10-23
   */
+@deprecated("This class will be removed in a future release", "v2.1")
 trait ManyClientDeviceUsersAccess
 	extends ManyRowModelAccess[ClientDeviceUser] with Indexed with FilterableView[ManyClientDeviceUsersAccess]
 {