+++ Citadel/src/utopia/citadel/database/model/description/CitadelDescriptionLinkModel.scala
@@ -14,6 +14,7 @@ object CitadelDescriptionLinkModel
 	/**
 	  * Database interaction model factory for ClientDevice description links
 	  */
+	@deprecated("This class will be removed in a future release", "v2.1")
 	lazy val clientDevice = DescriptionLinkModelFactory(CitadelTables.clientDeviceDescription)
 	/**
 	  * Database interaction model factory for DescriptionRole description links