+++ Metropolis/src/utopia/metropolis/model/stored/language/LanguageFamiliarity.scala
@@ -3,6 +3,8 @@ package utopia.metropolis.model.stored.language
 import utopia.flow.datastructure.immutable.Model
 import utopia.flow.generic.ValueConversions._
 import utopia.metropolis.model.StyledModelConvertible
+import utopia.metropolis.model.combined.description.LinkedDescription
+import utopia.metropolis.model.combined.language.DescribedLanguageFamiliarity
 import utopia.metropolis.model.partial.language.LanguageFamiliarityData
 import utopia.metropolis.model.stored.StoredModelConvertible
 
@@ -16,5 +18,16 @@ import utopia.metropolis.model.stored.StoredModelConvertible
 case class LanguageFamiliarity(id: Int, data: LanguageFamiliarityData) 
 	extends StoredModelConvertible[LanguageFamiliarityData] with StyledModelConvertible
 {
+	// IMPLEMENTED  ----------------------
+	
 	override def toSimpleModel = Model(Vector("id" -> id, "order_index" -> data.orderIndex))
+	
+	
+	// OTHER    --------------------------
+	
+	/**
+	  * @param descriptions Descriptions to apply to this language familiarity
+	  * @return A copy of this familiarity with those descriptions included
+	  */
+	def withDescriptions(descriptions: Set[LinkedDescription]) = DescribedLanguageFamiliarity(this, descriptions)
 }